requirements.txt
board
rpi_ws281x
adafruit-circuitpython-neopixel
web.py
